Move to google/cloud/internal.
Thanks to @mbrukman for pointing out we do not want this as
a public API.
