+++ google/cloud/CMakeLists.txt
@@ -65,14 +65,14 @@ add_library(google_cloud_cpp_common
     ${CMAKE_CURRENT_BINARY_DIR}/internal/build_info.cc
     internal/optional.h
     internal/port_platform.h
+    internal/random.h
+    internal/random.cc
     internal/setenv.h
     internal/setenv.cc
     internal/throw_delegate.h
     internal/throw_delegate.cc
     log.h
     log.cc
-    random.h
-    random.cc
     version.h)
 target_link_libraries(google_cloud_cpp_common PUBLIC Threads::Threads
     PRIVATE google_cloud_cpp_common_options)
@@ -89,8 +89,8 @@ google_cloud_cpp_add_clang_tidy(google_cloud_cpp_common)
 
 set(google_cloud_cpp_common_unit_tests
     internal/optional_test.cc
+    internal/random_test.cc
     internal/throw_delegate_test.cc
-    random_test.cc
     log_test.cc)
 
 # Export the list of unit tests so the Bazel BUILD file can pick it up.
