+++ google/cloud/bigtable/benchmarks/endurance_benchmark.cc
@@ -115,7 +115,7 @@ int main(int argc, char* argv[]) try {
 
 namespace {
 OperationResult RunOneApply(bigtable::Table& table, Benchmark const& benchmark,
-                            google::cloud::DefaultPRNG& generator) {
+                            google::cloud::internal::DefaultPRNG& generator) {
   auto row_key = benchmark.MakeRandomKey(generator);
   bigtable::SingleRowMutation mutation(std::move(row_key));
   for (int field = 0; field != kNumFields; ++field) {
@@ -127,7 +127,7 @@ OperationResult RunOneApply(bigtable::Table& table, Benchmark const& benchmark,
 
 OperationResult RunOneReadRow(bigtable::Table& table,
                               Benchmark const& benchmark,
-                              google::cloud::DefaultPRNG& generator) {
+                              google::cloud::internal::DefaultPRNG& generator) {
   auto row_key = benchmark.MakeRandomKey(generator);
   auto op = [&table, &row_key]() {
     auto row = table.ReadRow(
@@ -147,7 +147,7 @@ long RunBenchmark(bigtable::benchmarks::Benchmark& benchmark,
   auto data_client = benchmark.MakeDataClient();
   bigtable::Table table(std::move(data_client), table_id);
 
-  auto generator = google::cloud::MakeDefaultPRNG();
+  auto generator = google::cloud::internal::MakeDefaultPRNG();
 
   auto start = std::chrono::steady_clock::now();
   auto end = start + test_duration;
