+++ google/cloud/bigtable/benchmarks/random_mutation.cc
@@ -19,16 +19,17 @@ namespace google {
 namespace cloud {
 namespace bigtable {
 namespace benchmarks {
-bigtable::Mutation MakeRandomMutation(DefaultPRNG& gen, int f) {
+bigtable::Mutation MakeRandomMutation(google::cloud::internal::DefaultPRNG& gen,
+                                      int f) {
   std::string field = "field" + std::to_string(f);
   return bigtable::SetCell(kColumnFamily, std::move(field),
                            std::chrono::milliseconds(0), MakeRandomValue(gen));
 }
 
-std::string MakeRandomValue(DefaultPRNG& generator) {
+std::string MakeRandomValue(google::cloud::internal::DefaultPRNG& generator) {
   static std::string const letters(
       "ABCDEFGHIJLKMNOPQRSTUVWXYZabcdefghijlkmnopqrstuvwxyz0123456789-/_");
-  return google::cloud::Sample(generator, kFieldSize, letters);
+  return google::cloud::internal::Sample(generator, kFieldSize, letters);
 }
 }  // namespace benchmarks
 }  // namespace bigtable
