+++ google/cloud/bigtable/tests/admin_integration_test.cc
@@ -15,7 +15,7 @@
 #include "google/cloud/bigtable/grpc_error.h"
 #include "google/cloud/bigtable/internal/make_unique.h"
 #include "google/cloud/bigtable/testing/table_integration_test.h"
-#include "google/cloud/random.h"
+#include "google/cloud/internal/random.h"
 #include <gmock/gmock.h>
 #include <string>
 #include <vector>
