+++ google/cloud/google_cloud_cpp_common.bzl
@@ -3,17 +3,17 @@ google_cloud_cpp_common_HDRS = [
     "internal/build_info.h",
     "internal/optional.h",
     "internal/port_platform.h",
+    "internal/random.h",
     "internal/setenv.h",
     "internal/throw_delegate.h",
     "log.h",
-    "random.h",
     "version.h",
 ]
 
 google_cloud_cpp_common_SRCS = [
+    "internal/random.cc",
     "internal/setenv.cc",
     "internal/throw_delegate.cc",
     "log.cc",
-    "random.cc",
 ]
