+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -1,8 +1,8 @@
 # DO NOT EDIT -- GENERATED CODE
 google_cloud_cpp_common_unit_tests = [
     "internal/optional_test.cc",
+    "internal/random_test.cc",
     "internal/throw_delegate_test.cc",
-    "random_test.cc",
     "log_test.cc",
 ]
