+++ google/cloud/internal/random_test.cc
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/random.h"
+#include "random.h"
 #include <gmock/gmock.h>
 
-using namespace google::cloud;
+using namespace google::cloud::internal;
 
 TEST(BenchmarksRandom, Basic) {
   // This is not a statistical test for PRNG, basically we want to make
